+++ cmd/tagxl.go
@@ -68,7 +68,7 @@ var tagxlCmd = &cobra.Command{
 		}
 
 		ctx = context.WithValue(ctx, decoder.DEVEUI_CONTEXT_KEY, tagXlDevEui)
-		ctx = context.WithValue(ctx, decoder.PORT_CONTEXT_KEY, port)
+		ctx = context.WithValue(ctx, decoder.PORT_CONTEXT_KEY, uint8(port))
 		ctx = context.WithValue(ctx, decoder.FCNT_CONTEXT_KEY, 1) // Default frame count, can be adjusted as needed
 
 		data, err := d.Decode(ctx, args[1], uint8(port))
